refactor: split anilist/source; dedupe arr-auth middleware
- Extract anilist.rs rate-limit machinery into services/anilist/rate_limit.rs;
  parent re-exports anilist_cooldown_active and is_rate_limit_error so
  existing call sites still resolve at services::anilist::foo.
- Extract source.rs enums (Source, Resolution, WebKind, Origin, DecisionRule,
  SourceEvidence, ClassificationResult) into services/source/types.rs behind
  pub use.
- Collapse sonarr_compat::require_api_key and radarr_compat::require_api_key
  into a single handlers::arr_auth::check_api_key helper parameterised on the
  per-shim config selector and label. Each shim's wrapper is now three lines.

No behavior change. cargo test still 480 passes.
